new commit

+++ src/main/java/com/example/demoChat/repository/ordersRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ordersRepo extends JpaRepository<Order,Integer> {
+public interface ordersRepo extends JpaRepository<Order,Long> {
 }
